split out trapregisters class
problems with const

compiling

beforeclang

clang

singleton for TrapRegisters

partial structures

clang
